refactor: 게임 결과 팝업창 클래스 분리
